+++ Core/src/Polygon.cpp
@@ -41,14 +41,14 @@ void Polygon::AddVertex(const Vertex2& vertex)
 	m_vertices.push_back(vertex);
 }
 
-void Polygon::RemoveVertex(int index)
+void Polygon::RemoveVertex(size_t index)
 {
 	if (index >= m_vertices.size())
 		throw std::exception("Invalid Index");
 	m_vertices.erase(m_vertices.begin() + index, m_vertices.begin() + index + 1);
 }
 
-void Polygon::ReplaceVertex(int index, Vertex2 newVertex)
+void Polygon::ReplaceVertex(size_t index, Vertex2 newVertex)
 {
 	if (index >= m_vertices.size())
 		throw std::exception("Invalid Index");
